+++ src/components/view/mainPage/MainPage.ts
@@ -114,7 +114,6 @@ class MainPage {
       : `
         <div class="user_box_login user_box_link"><a href="/login" class="navigator">login</a></div>
         <div class="user_box_register user_box_link"><a href="/register" class="navigator">register</a></div>
-        <div class="user_box_profile user_box_link"><a href="/profile" class="navigator">profile</a></div>
       `;
   }
 }
