+++ src/index.ts
@@ -12,9 +12,9 @@ document.addEventListener('click', e => {
   if (element.tagName === 'A' && element.classList.contains('navigator')) {
     e.preventDefault();
     app.navigateTo(element.href);
-    alert(
-      'Рады вас видеть! Лучшие туры еще готовятся для вас. Просьба вернуться к нам через пару дней. Большое спасибо за понимание :)',
-    );
+    // alert(
+    //   'Рады вас видеть! Лучшие туры еще готовятся для вас. Просьба вернуться к нам через пару дней. Большое спасибо за понимание :)',
+    // );
   }
 });
 window.addEventListener('popstate', () => app.routerControllers());
